test: add use cases for pet

+++ apps/lifetails-backend/src/contexts/PetLifeMoments/application/remove/RemovePetLifeMomentUseCase.ts
@@ -1,3 +1,4 @@
+import { PetLifeMomentNotFoundException } from '../../domain/exceptions/PetLifeMomentNotFoundException';
 import { PetLifeMomentRepository } from '../../domain/repositories/PetLifeMomentRepository';
 import { RemovePetLifeMomentCommand } from './RemovePetLifeMomentCommand';
 
@@ -5,6 +6,11 @@ export class RemovePetLifeMomentUseCase {
   constructor(private readonly repository: PetLifeMomentRepository) {}
 
   async execute(command: RemovePetLifeMomentCommand): Promise<void> {
+    const petLifeMoment = await this.repository.find(command.id);
+    if (!petLifeMoment) {
+      throw new PetLifeMomentNotFoundException(command.id);
+    }
+
     await this.repository.remove(command.id);
   }
 }
